refactor: BaseLog를 기반으로 상속 구조 Log Domain refactor
