+++ src/main/java/yeah/yeahlogging/domain/document/ClickLogDocument.java
@@ -1,12 +1,14 @@
 package yeah.yeahlogging.domain.document;
 
-import lombok.*;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.Document;
 import org.springframework.data.elasticsearch.annotations.Field;
 import org.springframework.data.elasticsearch.annotations.FieldType;
 import org.springframework.data.elasticsearch.annotations.WriteTypeHint;
-import yeah.yeahlogging.domain.ClickLog;
+import yeah.yeahlogging.domain.log.ClickLog;
 
 import java.time.LocalDateTime;
 import java.util.Map;
