+++ src/main/java/yeah/yeahlogging/domain/document/ImpressionLogDocument.java
@@ -3,13 +3,12 @@ package yeah.yeahlogging.domain.document;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.NoArgsConstructor;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.elasticsearch.annotations.Document;
 import org.springframework.data.elasticsearch.annotations.Field;
 import org.springframework.data.elasticsearch.annotations.FieldType;
 import org.springframework.data.elasticsearch.annotations.WriteTypeHint;
-import yeah.yeahlogging.domain.ImpressionLog;
+import yeah.yeahlogging.domain.log.ImpressionLog;
 
 import java.time.LocalDateTime;
 import java.util.Map;
