+++ src/main/java/yeah/yeahlogging/domain/log/ClickLog.java
@@ -1,0 +1,11 @@
+package yeah.yeahlogging.domain.log;
+
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class ClickLog extends BaseLog {
+}
