+++ src/main/java/yeah/yeahlogging/domain/log/ScreenLog.java
@@ -1,0 +1,9 @@
+package yeah.yeahlogging.domain.log;
+
+import lombok.*;
+
+@Getter
+@Setter
+@NoArgsConstructor
+public class ScreenLog extends BaseLog {
+}
