+++ src/main/java/yeah/yeahlogging/service/LogService.java
@@ -4,9 +4,9 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.elasticsearch.core.ElasticsearchOperations;
 import org.springframework.stereotype.Service;
-import yeah.yeahlogging.domain.ClickLog;
-import yeah.yeahlogging.domain.ImpressionLog;
-import yeah.yeahlogging.domain.ScreenLog;
+import yeah.yeahlogging.domain.log.ClickLog;
+import yeah.yeahlogging.domain.log.ImpressionLog;
+import yeah.yeahlogging.domain.log.ScreenLog;
 import yeah.yeahlogging.domain.document.ClickLogDocument;
 import yeah.yeahlogging.domain.document.ImpressionLogDocument;
 import yeah.yeahlogging.domain.document.ScreenLogDocument;
